+++ packages/google-vertex/src/google-vertex-embedding-options.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod/v4';
+import * as z from 'zod/v4';
 
 // https://cloud.google.com/vertex-ai/generative-ai/docs/model-reference/text-embeddings-api
 export type GoogleVertexEmbeddingModelId =