+++ packages/openai-compatible/src/completion/openai-compatible-completion-language-model.ts
@@ -18,7 +18,7 @@ import {
   postJsonToApi,
   ResponseHandler,
 } from '@ai-sdk/provider-utils';
-import { z } from 'zod/v4';
+import * as z from 'zod/v4';
 import {
   defaultOpenAICompatibleErrorStructure,
   ProviderErrorStructure,