+++ packages/provider-utils/src/validator.ts
@@ -50,10 +50,51 @@ export function isValidator(value: unknown): value is Validator {
   );
 }
 
+/**
+ * Creates a validator with deferred creation.
+ * This is important to reduce the startup time of the library
+ * and to avoid initializing unused validators.
+ *
+ * @param createValidator A function that creates a validator.
+ * @returns A function that returns a validator.
+ */
+export function lazyValidator<OBJECT>(
+  createValidator: () => Validator<OBJECT>,
+): LazyValidator<OBJECT> {
+  // cache the validator to avoid initializing it multiple times
+  let validator: Validator<OBJECT> | undefined;
+  return () => {
+    if (validator == null) {
+      validator = createValidator();
+    }
+    return validator;
+  };
+}
+
+export type LazyValidator<OBJECT> = () => Validator<OBJECT>;
+
+export type FlexibleValidator<OBJECT> =
+  | Validator<OBJECT>
+  | LazyValidator<OBJECT>
+  | StandardSchemaV1<unknown, OBJECT>;
+
+export type InferValidator<SCHEMA> =
+  SCHEMA extends StandardSchemaV1<unknown, infer T>
+    ? T
+    : SCHEMA extends LazyValidator<infer T>
+      ? T
+      : SCHEMA extends Validator<infer T>
+        ? T
+        : never;
+
 export function asValidator<OBJECT>(
-  value: Validator<OBJECT> | StandardSchemaV1<unknown, OBJECT>,
+  value: FlexibleValidator<OBJECT>,
 ): Validator<OBJECT> {
-  return isValidator(value) ? value : standardSchemaValidator(value);
+  return isValidator(value)
+    ? value
+    : typeof value === 'function'
+      ? value()
+      : standardSchemaValidator(value);
 }
 
 export function standardSchemaValidator<OBJECT>(