+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.0-beta.18
+
+### Patch Changes
+
+- 95f65c2: chore: use import \* from zod/v4
+- Updated dependencies [95f65c2]
+- Updated dependencies [95f65c2]
+  - @ai-sdk/provider-utils@4.0.0-beta.11
+  - @ai-sdk/anthropic@3.0.0-beta.15
+
 ## 4.0.0-beta.17
 
 ### Major Changes