+++ packages/elevenlabs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/elevenlabs
 
+## 2.0.0-beta.11
+
+### Patch Changes
+
+- 95f65c2: chore: use import \* from zod/v4
+- Updated dependencies [95f65c2]
+- Updated dependencies [95f65c2]
+  - @ai-sdk/provider-utils@4.0.0-beta.11
+
 ## 2.0.0-beta.10
 
 ### Major Changes